move migration code to javamigration package and add utils package
