+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib/src/main/java/com/microsoft/azure/toolkit/intellij/connector/projectexplorer/MigrateToAzureFacetNode.java
@@ -10,13 +10,13 @@ import com.intellij.ide.projectView.PresentationData;
 import com.intellij.openapi.extensions.ExtensionPointName;
 import com.intellij.openapi.project.Project;
 import com.intellij.ui.tree.LeafState;
-import com.microsoft.azure.toolkit.intellij.appmod.AppModPanelHelper;
-import com.microsoft.azure.toolkit.intellij.appmod.AppModUtils;
-import com.microsoft.azure.toolkit.intellij.appmod.Constants;
+import com.microsoft.azure.toolkit.intellij.appmod.utils.AppModPanelHelper;
+import com.microsoft.azure.toolkit.intellij.appmod.utils.AppModUtils;
+import com.microsoft.azure.toolkit.intellij.appmod.utils.Constants;
 import com.microsoft.azure.toolkit.intellij.common.IntelliJAzureIcons;
 import com.microsoft.azure.toolkit.intellij.connector.dotazure.AzureModule;
-import com.microsoft.azure.toolkit.intellij.appmod.IMigrateOptionProvider;
-import com.microsoft.azure.toolkit.intellij.appmod.MigrateNodeData;
+import com.microsoft.azure.toolkit.intellij.appmod.javamigration.IMigrateOptionProvider;
+import com.microsoft.azure.toolkit.intellij.appmod.javamigration.MigrateNodeData;
 import com.microsoft.azure.toolkit.intellij.appmod.common.AppModPluginInstaller;
 import com.microsoft.azure.toolkit.ide.common.icon.AzureIcons;
 import com.microsoft.azure.toolkit.lib.common.action.Action;
